backend/requirements.txt
fastapi
uvicorn
opencv-python
pydantic
numpy
onnx
onnxruntime-gpu
scikit-image
moviepy==1.0.3
python-multipart